homework_13/requirements.txt
fastapi
uvicorn
transformers
PyYAML
torch
numpy==1.24.1
requests
python-telegram-bot==21.6
